move utils to common files
